refactor: move fetch roles to roles store and fetch on successful crud
